15강 PUT, DELETE 에 대한 예외추가 - UserController에 PUT, DELETE 예외 추가 - 추가적인 파일 생성 없음 - 한 클래스인 Controller에 많은 기능이 들어가 있다는 문제가 많다.  - 나중가면은 굉장히 길어져 코드를 읽기 어려워진다.

+++ src/main/java/com/group/libraryapp/controller/user/UserController.java
@@ -22,29 +22,14 @@ public class UserController {
 
     @PostMapping("/user") // POST /user
     public void saveUser(@RequestBody UserCreateRequest request) {
-        String sql = "INSERT INTO user (name, age) VALUES (?, ?)"; // ?는 문자열 변수를 의미
-        jdbcTemplate.update(sql, request.getName(), request.getAge()); //sql을 문장으로 쓰고, 첫번째 ?에는 getName, 두번째 ?에는 getAge를 넣어준다.
+        String sql = "INSERT INTO user (name, age) VALUES (?, ?)";
+        jdbcTemplate.update(sql, request.getName(), request.getAge());
     }
 
     @GetMapping("/user")
     public List<UserResponse> getUsers() {
         String sql = "SELECT * FROM user";
-        /*
-        return jdbcTemplate.query(sql, new RowMapper<UserResponse>() {
-            //Crtl + O 를 눌러준다. -> MapRow를 오버라이드 해준다.
-            // map.row는 유저 정보를 선언한 타입인 userResponse로 바꿔준다.
-            @Override
-            public UserResponse mapRow(ResultSet rs, int rowNum) throws SQLException {
-                long id = rs.getLong("id");
-                String name = rs.getString("name"); //ResultSet인 rs에서 string 타입의 name을 가져온다.
-                int age = rs.getInt("age");
-                return new UserResponse(id,name,age);
-            }
-        });
-        */
 
-        //람다식 - RowMapper 에다가 람다식 전환
-        // 람다를 잘모른다면 자발개발자의 코틀린 입문 강의의 람다식 미리보기 강의 확인해볼 것(20분짜리)
         return jdbcTemplate.query(sql, (rs, rowNum) -> {
             long id = rs.getLong("id");
             String name = rs.getString("name");
@@ -55,6 +40,12 @@ public class UserController {
 
     @PutMapping("/user")
     public void updateUser(@RequestBody UserUpdateRequest request) {
+        String readSql = "SELECT * FROM user WHERE id = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, request.getId()).isEmpty(); //찾는 게 있다면, 리스트에 0이 담기지만(true), 아닐 시 null이 된다.
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "UPDATE user SET name = ? WHERE id = ?";
 
         jdbcTemplate.update(sql, request.getName(), request.getId());
@@ -63,6 +54,12 @@ public class UserController {
 
     @DeleteMapping("/user")
     public void deleteUser(@RequestParam String name) {
+        String readSql = "SELECT * FROM user WHERE name = ?";
+        boolean isUserNotExist = jdbcTemplate.query(readSql, (rs, rowNum) -> 0, name).isEmpty(); //찾는 게 있다면, 리스트에 0이 담기지만(true), 아닐 시 null이 된다.
+        if (isUserNotExist) {
+            throw new IllegalArgumentException();
+        }
+
         String sql = "DELETE FROM user WHERE name = ?";
         jdbcTemplate.update(sql, name);
     }
